chore: move urlMatch to isomorphic
To be reused in injected code.
