+++ packages/playwright-core/src/client/frame.ts
@@ -30,9 +30,9 @@ import type { Page } from './page';
 import { EventEmitter } from 'events';
 import { Waiter } from './waiter';
 import { Events } from './events';
-import type { LifecycleEvent, URLMatch, SelectOption, SelectOptionOptions, FilePayload, WaitForFunctionOptions, StrictOptions } from './types';
+import type { LifecycleEvent, SelectOption, SelectOptionOptions, FilePayload, WaitForFunctionOptions, StrictOptions } from './types';
 import { kLifecycleEvents } from './types';
-import { urlMatches } from '../utils/network';
+import { type URLMatch, urlMatches } from '../utils';
 import type * as api from '../../types/types';
 import type * as structs from '../../types/structs';
 import { addSourceUrlToScript } from './clientHelper';
