+++ packages/playwright-core/src/client/harRouter.ts
@@ -18,7 +18,7 @@ import { debugLogger } from '../utils/debugLogger';
 import type { BrowserContext } from './browserContext';
 import type { LocalUtils } from './localUtils';
 import type { Route } from './network';
-import type { URLMatch } from './types';
+import type { URLMatch } from '../utils';
 import type { Page } from './page';
 
 type HarNotFoundAction = 'abort' | 'fallback';
