+++ packages/playwright-core/src/client/network.ts
@@ -28,9 +28,8 @@ import { Events } from './events';
 import type { Page } from './page';
 import { Waiter } from './waiter';
 import type * as api from '../../types/types';
-import type { HeadersArray, URLMatch } from '../common/types';
-import { urlMatches } from '../utils/network';
-import { MultiMap } from '../utils/multimap';
+import type { HeadersArray } from '../common/types';
+import { MultiMap, urlMatches, type URLMatch } from '../utils';
 import { APIResponse } from './fetch';
 import type { Serializable } from '../../types/structs';
 import type { BrowserContext } from './browserContext';
