+++ packages/playwright-core/src/server/playwright.ts
@@ -24,10 +24,9 @@ import { Selectors } from './selectors';
 import { WebKit } from './webkit/webkit';
 import type { CallMetadata } from './instrumentation';
 import { createInstrumentation, SdkObject } from './instrumentation';
-import { debugLogger } from '../utils/debugLogger';
+import { debugLogger, type Language } from '../utils';
 import type { Page } from './page';
 import { DebugController } from './debugController';
-import type { Language } from '../utils/isomorphic/locatorGenerators';
 import type { BrowserType } from './browserType';
 
 type PlaywrightOptions = {
