+++ packages/playwright-core/src/server/recorder/jsonl.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { asLocator } from '../../utils/isomorphic/locatorGenerators';
+import { asLocator } from '../../utils';
 import type { ActionInContext } from './codeGenerator';
 import type { Language, LanguageGenerator, LanguageGeneratorOptions } from './language';
