+++ packages/playwright-core/src/server/types.ts
@@ -16,7 +16,7 @@
  */
 
 import type { Size, Point, TimeoutOptions, HeadersArray } from '../common/types';
-export type { Size, Point, Rect, Quad, URLMatch, TimeoutOptions, HeadersArray } from '../common/types';
+export type { Size, Point, Rect, Quad, TimeoutOptions, HeadersArray } from '../common/types';
 import type * as channels from '@protocol/channels';
 
 export type StrictOptions = {
