+++ packages/playwright-core/src/utils/index.ts
@@ -18,15 +18,18 @@ export * from './ascii';
 export * from './comparators';
 export * from './crypto';
 export * from './debug';
+export * from './debugLogger';
 export * from './env';
 export * from './eventsHelper';
 export * from './fileUtils';
-export * from './glob';
 export * from './headers';
 export * from './hostPlatform';
 export * from './httpServer';
 export * from './manualPromise';
+export * from './isomorphic/locatorGenerators';
 export * from './isomorphic/mimeType';
+export * from './isomorphic/stringUtils';
+export * from './isomorphic/urlMatch';
 export * from './multimap';
 export * from './network';
 export * from './processLauncher';
@@ -43,5 +46,3 @@ export * from './userAgent';
 export * from './wsServer';
 export * from './zipFile';
 export * from './zones';
-export * from './isomorphic/locatorGenerators';
-export * from './isomorphic/stringUtils';
