+++ packages/playwright-core/src/utils/isomorphic/urlMatch.ts
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+import { isString } from './stringUtils';
+
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_expressions#escaping
 const escapedChars = new Set(['$', '^', '+', '.', '*', '(', ')', '|', '\\', '?', '{', '}', '[', ']']);
 
@@ -78,4 +80,53 @@ export function globToRegex(glob: string): RegExp {
   }
   tokens.push('$');
   return new RegExp(tokens.join(''));
+}
+
+function isRegExp(obj: any): obj is RegExp {
+  return obj instanceof RegExp || Object.prototype.toString.call(obj) === '[object RegExp]';
+}
+
+export type URLMatch = string | RegExp | ((url: URL) => boolean);
+
+export function urlMatchesEqual(match1: URLMatch, match2: URLMatch) {
+  if (isRegExp(match1) && isRegExp(match2))
+    return match1.source === match2.source && match1.flags === match2.flags;
+  return match1 === match2;
+}
+
+export function urlMatches(baseURL: string | undefined, urlString: string, match: URLMatch | undefined): boolean {
+  if (match === undefined || match === '')
+    return true;
+  if (isString(match) && !match.startsWith('*'))
+    match = constructURLBasedOnBaseURL(baseURL, match);
+  if (isString(match))
+    match = globToRegex(match);
+  if (isRegExp(match))
+    return match.test(urlString);
+  if (typeof match === 'string' && match === urlString)
+    return true;
+  const url = parsedURL(urlString);
+  if (!url)
+    return false;
+  if (typeof match === 'string')
+    return url.pathname === match;
+  if (typeof match !== 'function')
+    throw new Error('url parameter should be string, RegExp or function');
+  return match(url);
+}
+
+function parsedURL(url: string): URL | null {
+  try {
+    return new URL(url);
+  } catch (e) {
+    return null;
+  }
+}
+
+export function constructURLBasedOnBaseURL(baseURL: string | undefined, givenURL: string): string {
+  try {
+    return (new URL(givenURL, baseURL)).toString();
+  } catch (e) {
+    return givenURL;
+  }
 }
