+++ packages/playwright-core/src/utils/network.ts
@@ -22,9 +22,6 @@ import type net from 'net';
 import { getProxyForUrl } from '../utilsBundle';
 import { HttpsProxyAgent } from '../utilsBundle';
 import url from 'url';
-import type { URLMatch } from '../common/types';
-import { isString, isRegExp } from './rtti';
-import { globToRegex } from './glob';
 import { httpHappyEyeballsAgent, httpsHappyEyeballsAgent } from './happy-eyeballs';
 
 export type HTTPRequestParams = {
@@ -109,49 +106,6 @@ export function fetchData(params: HTTPRequestParams, onError?: (params: HTTPRequ
       response.on('end', () => resolve(body));
     }, reject);
   });
-}
-
-export function urlMatchesEqual(match1: URLMatch, match2: URLMatch) {
-  if (isRegExp(match1) && isRegExp(match2))
-    return match1.source === match2.source && match1.flags === match2.flags;
-  return match1 === match2;
-}
-
-export function urlMatches(baseURL: string | undefined, urlString: string, match: URLMatch | undefined): boolean {
-  if (match === undefined || match === '')
-    return true;
-  if (isString(match) && !match.startsWith('*'))
-    match = constructURLBasedOnBaseURL(baseURL, match);
-  if (isString(match))
-    match = globToRegex(match);
-  if (isRegExp(match))
-    return match.test(urlString);
-  if (typeof match === 'string' && match === urlString)
-    return true;
-  const url = parsedURL(urlString);
-  if (!url)
-    return false;
-  if (typeof match === 'string')
-    return url.pathname === match;
-  if (typeof match !== 'function')
-    throw new Error('url parameter should be string, RegExp or function');
-  return match(url);
-}
-
-function parsedURL(url: string): URL | null {
-  try {
-    return new URL(url);
-  } catch (e) {
-    return null;
-  }
-}
-
-export function constructURLBasedOnBaseURL(baseURL: string | undefined, givenURL: string): string {
-  try {
-    return (new URL(givenURL, baseURL)).toString();
-  } catch (e) {
-    return givenURL;
-  }
 }
 
 export function createHttpServer(requestListener?: (req: http.IncomingMessage, res: http.ServerResponse) => void): http.Server;
